🔧[Chore]#5 - Enum 파일 분리, 오타 수정
