+++ Muje/Muje/Common/Enum/ApplicationManagement/ApplicationTab.swift
